requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
aiohttp==3.9.1
python-dotenv==1.0.0

# PyTorch and CUDA support
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image generation
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0

# Text-to-speech
chatterbox-tts>=0.1.0

# Text generation
tokenizers>=0.19.0

# Additional dependencies
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0
tqdm>=4.65.0
